backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
playwright==1.40.0
python-multipart==0.0.6
aiosmtplib==3.0.1
email-validator==2.1.0
